# front/src/app/services/models/offre-response.ts
/* tslint:disable */
/* eslint-disable */
export interface OffreResponse {
  archived?: boolean;
  dateExpiration?: string;
  datePublication?: string;
  description?: string;
  id?: number;
  profil?: string;
  publisher?: string;
  responsabilites?: string;
  shareable?: boolean;
  title?: string;
  typeContrat?: 'Cdi' | 'Stage' | 'cdd';
}