# front/src/openApi/openApi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "OpenApi specification - Alibou",
        "description": "OpenApi documentation for Spring Security",
        "termsOfService": "Terms of service",
        "contact": {
            "name": "Alibou",
            "url": "https://aliboucoding.com/course",
            "email": "contact@aliboucoding.com"
        },
        "license": {
            "name": "Licence name",
            "url": "https://some-url.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8088",
            "description": "Local ENV"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "paths": {
...
        "/offre/update/{id}": {
            "post": {
                "tags": [
                    "offre-controller"
                ],
                "operationId": "updateOffre",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OffreRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/OffreResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/offre/addFavorite": {
            "post": {
                "tags": [
                    "offre-controller"
                ],
                "operationId": "addFavorite",
                "parameters": [
                    {
                        "name": "offerId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/offre/AddOffre": {
            "post": {
                "tags": [
                    "offre-controller"
                ],
                "operationId": "saveBook",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OffreRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "integer",
                                    "format": "int32"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/offre/archived/{offer-id}": {
            "patch": {
                "tags": [
                    "offre-controller"
                ],
                "operationId": "updateArchivedStatus",
                "parameters": [
                    {
                        "name": "offer-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "integer",
                                    "format": "int32"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/users/favorite-offers/{email}": {
            "get": {
                "tags": [
                    "user-controler"
                ],
                "operationId": "getUserFavoriteOffers",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "uniqueItems": true,
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/OffreResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/offre/{id}": {
            "get": {
                "tags": [
                    "offre-controller"
                ],
                "operationId": "getOffreById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/OffreResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/users/favorites/{email}/{offerId}": {
            "delete": {
                "tags": [
                    "user-controler"
                ],
                "operationId": "removeFavoriteOffer",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "offerId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/offre/delete/{id}": {
            "delete": {
                "tags": [
                    "offre-controller"
                ],
                "operationId": "deleteOffre",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "OffreRequest": {
                "required": [
                    "description",
                    "expirationDate",
                    "profil",
                    "responsabilites",
                    "title"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "responsabilites": {
                        "type": "string"
                    },
                    "profil": {
                        "type": "string"
                    },
                    "expirationDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "shareable": {
                        "type": "boolean"
                    },
                    "typeContrat": {
                        "type": "string",
                        "enum": [
                            "Cdi",
                            "Stage",
                            "cdd"
                        ]
                    },
                    "publisher": {
                        "type": "string"
                    }
                }
            },
            "OffreResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "responsabilites": {
                        "type": "string"
                    },
                    "profil": {
                        "type": "string"
                    },
                    "datePublication": {
                        "type": "string",
                        "format": "date"
                    },
                    "dateExpiration": {
                        "type": "string",
                        "format": "date"
                    },
                    "typeContrat": {
                        "type": "string",
                        "enum": [
                            "Cdi",
                            "Stage",
                            "cdd"
                        ]
                    },
                    "archived": {
                        "type": "boolean"
                    },
                    "shareable": {
                        "type": "boolean"
                    },
                    "publisher": {
                        "type": "string"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT auth description",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}